requirement.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
importlib-metadata==0.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.18.0
nose==1.3.7
path.py==11.5.0
requests==2.21.0
selenium==3.141.0
urllib3==1.24.1
webargs==5.1.2
Werkzeug==0.14.1
zipp==0.3.3
